+++ src/tools/math.cc
@@ -12,7 +12,7 @@ namespace math{
 int add1(int x, int y)
 {
   int count = 12;
-  for (size_t i = 0; i < count; i++) {
+  for (int i = 0; i < count; i++) {
     std::this_thread::sleep_for(std::chrono::seconds(1));
   }
   return x + y;
@@ -20,7 +20,7 @@ int add1(int x, int y)
 int add2(int x, int y)
 {
   int count = 7;
-  for (size_t i = 0; i < count; i++) {
+  for (int i = 0; i < count; i++) {
     std::this_thread::sleep_for(std::chrono::seconds(1));
   }
   return x + y;
@@ -28,7 +28,7 @@ int add2(int x, int y)
 int add3(int x, int y)
 {
   int count = 5;
-  for (size_t i = 0; i < count; i++) {
+  for (int i = 0; i < count; i++) {
     std::this_thread::sleep_for(std::chrono::seconds(1));
   }
   return x + y;
